Fix: Rename formatDateByYearMonthDays to formatCreatedAtDate
