+++ apps/web/components/user-feedback/dynamic-feedback-table.tsx
@@ -9,7 +9,7 @@ import {
   TableRow,
 } from "@repo/ui/components/table";
 import FeedbackBadge from "@/components/user-feedback/feedback-badge";
-import { formatCreatedAtDate } from "@/utils/dateUtils";
+import { formatCreatedAtDate } from "@/utils/date-utils";
 import SkeletonFeedbackItem from "@/components/user-feedback/skeleton-feedback-item";
 import NoFeedbackMessage from "@/components/user-feedback/no-feedback-message";
 import { SentimentAnalysisResult } from "@/types/sentimentAnalysisResult";
